services/ingest/requirements.txt
pika==1.3.2
opencv-python==4.8.1.78
pyyaml==6.0
onvif2-zeep==0.8.1
WSDiscovery==0.9.0